Speedy CLI due to lazy imports #40
