+++ src/tests/test_inference.py
@@ -1,13 +1,14 @@
 
-from ftw_cli.inference import create_input, polygonize, run
 from click.testing import CliRunner
-import pytest
+
+from ftw_cli.cli import inference_download, inference_polygonize, inference_run
+
 
 def test_inference_download(): # create_input
     runner = CliRunner()
 
     # Check help
-    result = runner.invoke(create_input, ["--help"])
+    result = runner.invoke(inference_download, ["--help"])
     assert result.exit_code == 0, result.output
     assert "Usage: download [OPTIONS]" in result.output
 
@@ -17,7 +18,7 @@ def test_inference_polygonize():
     runner = CliRunner()
 
     # Check help
-    result = runner.invoke(polygonize, ["--help"])
+    result = runner.invoke(inference_polygonize, ["--help"])
     assert result.exit_code == 0, result.output
     assert "Usage: polygonize [OPTIONS] INPUT" in result.output
 
@@ -27,7 +28,7 @@ def test_inference_run():
     runner = CliRunner()
 
     # Check help
-    result = runner.invoke(run, ["--help"])
+    result = runner.invoke(inference_run, ["--help"])
     assert result.exit_code == 0, result.output
     assert "Usage: run [OPTIONS] INPUT" in result.output
