+++ src/tests/test_model.py
@@ -1,13 +1,14 @@
 
-from ftw_cli.model import fit, test
 from click.testing import CliRunner
-import pytest
+
+from ftw_cli.cli import model_fit, model_test
+
 
 def test_model_fit():
     runner = CliRunner()
 
     # Check help
-    result = runner.invoke(fit, ["--help"])
+    result = runner.invoke(model_fit, ["--help"])
     assert result.exit_code == 0, result.output
     assert "Usage: fit [OPTIONS] [CLI_ARGS]..." in result.output
 
@@ -17,7 +18,7 @@ def test_model_test():
     runner = CliRunner()
 
     # Check help
-    result = runner.invoke(test, ["--help"])
+    result = runner.invoke(model_test, ["--help"])
     assert result.exit_code == 0, result.output
     assert "Usage: test [OPTIONS] [CLI_ARGS]..." in result.output
